Extract PageTitle to separate component
