+++ src/extensions/query-translator/component/QueryTranslator.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { SideNavigationItem } from '@neo4j-ndl/react';
+import { MenuItem } from '@neo4j-ndl/react';
 import QueryTranslatorSettingsModal from './QueryTranslatorSettingsModal';
-import { Tooltip } from '@mui/material';
 import { ExclamationTriangleIconSolid, LanguageIconSolid } from '@neo4j-ndl/react/icons';
 import { getModelProvider } from '../state/QueryTranslatorSelector';
 
@@ -10,33 +9,24 @@ const QueryTranslatorButton = (active) => {
   const [open, setOpen] = React.useState(false);
   const button = (
     <div>
-      <Tooltip title={'Natural Language Queries'} aria-label='examples'>
-        <SideNavigationItem
-          onClick={() => setOpen(true)}
-          icon={
-            <>
-              <LanguageIconSolid
-
-              // className={navItemClass}
-              />
-              {/* TODO Use Needle Icon Badges when implemented. */}
-              {active.active == '' || active.active == undefined ? (
-                <ExclamationTriangleIconSolid
-                  color='red'
-                  style={{
-                    marginTop: '-26px',
-                    marginLeft: '11px',
-                  }}
-                ></ExclamationTriangleIconSolid>
-              ) : (
-                <></>
-              )}
-            </>
-          }
-        >
-          Natural Language Queries
-        </SideNavigationItem>
-      </Tooltip>
+      <MenuItem
+        title='Natural Language Queries'
+        onClick={() => setOpen(true)}
+        icon={
+          <>
+            <LanguageIconSolid />
+            {/* TODO Use Needle Icon Badges when implemented. */}
+            {active.active == '' || active.active == undefined ? (
+              <ExclamationTriangleIconSolid
+                color='red'
+                className='-n-mt-1 n-ml-2 n-w-4/5'
+              ></ExclamationTriangleIconSolid>
+            ) : (
+              <></>
+            )}
+          </>
+        }
+      />
     </div>
   );
